Data imported, visualized and analysed
The train, validation and test data was imported to the project,
visualized and analysed. In order to do that, 100 images from the three
datasets where printed, so it can be seen how they vary in lightness,
sharpness and other properties. A histogram was created in order to see
which images are more frequent.