Model worked on web images
Model now works with new images. The problem was TensorFlow not being
able to recover the saved model because not all the variables,
placeholders and tensors were declared and it did not know where to
put the saved data from the recovery file into the python environment.
This is an issue to remember for the future, because it was very tricky
to find out what was the problem.
The results on the image where analysed.